# operator-api/openapi.yaml
openapi: 3.1.0
info:
  title: Kernel Images API
  version: 0.1.0
paths:
...
  /computer/click_mouse:
    post:
      summary: Click mouse (compat)
      operationId: computerClickMouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClickMouseRequest"
            examples:
              default:
                value: { button: left, click_type: click, x: 640, y: 360, num_clicks: 1 }
      responses:
        "200":
          description: Clicked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalError"
  /computer/move_mouse:
    post:
      summary: Move mouse (compat)
      operationId: computerMoveMouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveMouseRequest"
            examples:
              default:
                value: { x: 800, y: 450 }
      responses:
        "200":
          description: Moved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalError"

  # -------------------------
  # NEW ENDPOINTS (ADDITIVE)
  # -------------------------
...
  /input/mouse/move:
    post:
      summary: Move mouse cursor
      operationId: inputMouseMove
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MouseMoveRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/mouse/click:
    post:
      summary: Click mouse button
      operationId: inputMouseClick
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MouseClickRequest" }
            examples:
              default:
                value: { button: left, count: 1 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/mouse/scroll:
    post:
      summary: Scroll mouse wheel
      operationId: inputMouseScroll
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MouseScrollRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/keyboard/type:
    post:
      summary: Type text (IME-aware)
      operationId: inputKeyboardType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KeyboardTypeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/keyboard/key_down:
    post:
      summary: Key down
      operationId: inputKeyDown
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KeyRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/keyboard/key_up:
    post:
      summary: Key up
      operationId: inputKeyUp
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KeyRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/window/activate:
    post:
      summary: Activate window by match
      operationId: inputWindowActivate
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowActivateRequest" }
      responses:
        "200":
          description: Activated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WindowActivateResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/window/move_resize:
    post:
      summary: Move and resize window
      operationId: inputWindowMoveResize
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowMoveResizeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/window/close:
    post:
      summary: Close window by match
      operationId: inputWindowClose
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowCloseRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/mouse/move_relative:
    post:
      summary: Move mouse cursor relative
      operationId: inputMouseMoveRelative
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MouseMoveRelativeRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OkResponse" }
        "400":
          $ref: "#/components/responses/BadRequestError"
  /input/mouse/down:
    post:
      summary: Mouse button down
      operationId: inputMouseDown
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MouseButtonRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequestError" }
  /input/mouse/up:
    post:
      summary: Mouse button up
      operationId: inputMouseUp
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MouseButtonRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequestError" }
...
  /input/keyboard/key:
    post:
      summary: Send one or more key strokes
      operationId: inputKeyboardKey
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/KeyboardKeysRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
        "400": { $ref: "#/components/responses/BadRequestError" }
  /input/window/focus:
    post:
      summary: Focus window by match
      operationId: inputWindowFocus
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowActivateRequest" }
      responses:
        "200":
          description: Focused
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WindowFocusResponse" }
        "400": { $ref: "#/components/responses/BadRequestError" }
  /input/window/raise:
    post:
      summary: Raise window
      operationId: inputWindowRaise
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowActivateRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
  /input/window/minimize:
    post:
      summary: Minimize window
      operationId: inputWindowMinimize
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowActivateRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
  /input/window/map:
    post:
      summary: Map (show) window
      operationId: inputWindowMap
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowActivateRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
  /input/window/unmap:
    post:
      summary: Unmap (hide) window
      operationId: inputWindowUnmap
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowActivateRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
  /input/window/kill:
    post:
      summary: Kill window
      operationId: inputWindowKill
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowActivateRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
...
  /input/window/name:
    post:
      summary: Get window name
      operationId: inputWindowName
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowIdRequest" }
      responses:
        "200": { description: Name, content: { application/json: { schema: { $ref: "#/components/schemas/WindowNameResponse" } } } }
  /input/window/pid:
    post:
      summary: Get window pid
      operationId: inputWindowPid
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowIdRequest" }
      responses:
        "200": { description: PID, content: { application/json: { schema: { $ref: "#/components/schemas/WindowPidResponse" } } } }
  /input/window/geometry:
    post:
      summary: Get window geometry
      operationId: inputWindowGeometry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowIdRequest" }
      responses:
        "200": { description: Geometry, content: { application/json: { schema: { $ref: "#/components/schemas/WindowGeometryResponse" } } } }
...
  /input/desktop/count:
    get:
      summary: Get number of desktops
      operationId: inputDesktopCountGet
      responses:
        "200": { description: Count, content: { application/json: { schema: { $ref: "#/components/schemas/DesktopCountResponse" } } } }
    post:
      summary: Set number of desktops
      operationId: inputDesktopCountSet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DesktopCountRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
  /input/desktop/current:
    get:
      summary: Get current desktop index
      operationId: inputDesktopCurrentGet
      responses:
        "200": { description: Index, content: { application/json: { schema: { $ref: "#/components/schemas/DesktopIndexResponse" } } } }
    post:
      summary: Set current desktop index
      operationId: inputDesktopCurrentSet
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DesktopIndexRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
  /input/desktop/window_desktop:
    post:
      summary: Move window to desktop
      operationId: inputDesktopWindowDesktop
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WindowDesktopRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/OkResponse" } } } }
...
  /input/combo/activate_and_type:
    post:
      summary: Activate a window then type
      operationId: comboActivateAndType
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ActivateAndTypeRequest" }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ComboWindowResponse" } } } }
...
components:
  schemas:
...
    ClickMouseRequest:
      type: object
      properties:
        button:
          type: string
          description: Mouse button to interact with
          enum: [left, right, middle, back, forward]
        click_type:
          type: string
          description: Type of click action
          enum: [down, up, click]
        x:
          type: integer
          description: X coordinate of the click position
        y:
          type: integer
          description: Y coordinate of the click position
        hold_keys:
          type: array
          description: Modifier keys to hold during the click
          items:
            type: string
        num_clicks:
          type: integer
          description: Number of times to repeat the click
          default: 1
      additionalProperties: false
...
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
...
    MouseClickRequest:
      type: object
      properties:
        button:
          oneOf:
            - type: string
              enum: [left, middle, right, back, forward]
            - type: integer
        count: { type: integer, default: 1 }
      required: [button]
...
    KeyRequest:
      type: object
      properties:
        key: { type: string }
      required: [key]
...
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"